Refactor Entity::Kill to use shooting mechanism
The `Kill` method in the `Entity` class has been updated to remove the previous health-setting logic. Instead of setting the entity's health to zero or triggering a weapon damage event, it now retrieves the current weapon hash and uses the `MISC::SHOOT_SINGLE_BULLET_BETWEEN_COORDS` function to simulate a shooting action around the entity's position.